web-animations/animation-player/animation-player-reverse.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player reverse() method test</title>
<meta name="assert" content="Inverts the playback rate of this player and seeks to the start of the source content if it has finished playing in the reversed direction using the reverse a player procedure for this object">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationplayer-reverse">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
// TODO check that assert text is according to the spec

test(function() {
    var player = new AnimationPlayer(createDiv(this));
    assert_throws('InvalidStateError', function() {
        player.reverse();
    }, 'AnimationPlayer.reverse() should throw InvalidStateError if player has no ' +
        'associated timeline');
}, 'Test that AnimationPlayer.reverse() throws InvalidStateError if player has no ' +
    'associated timeline');

test(function() {
    var timeline = new InactiveTimeline();
    var p = timeline.play(createDiv(this));
    var player = new AnimationPlayer(createDiv(this), timeline);
    assert_throws('InvalidStateError', function() {
        player.reverse();
    }, 'AnimationPlayer.reverse() should throw InvalidStateError if timeline, associated ' +
        'with the player is inactive');
}, 'Test that AnimationPlayer.reverse() throws InvalidStateError if player\'s ' +
    'associated timeline is inactive');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.reverse();

    assert_equals(player.playbackRate, -1, 'Playback rate should be changed in sign');
}, 'Test that AnimationPlayer.reverse() changes playback rate in sign');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -2;
    player.reverse();

    assert_equals(player.playbackRate, 2, 'Playback rate should be changed in sign');
}, 'Test that AnimationPlayer.reverse() changes playback rate in sign (from negative to positive)');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.currentTime = 100;
    player.pause();
    player.reverse();

    player.ready.then(t.step_func(function() {
        assert_equals(player.playbackRate, -1, 'Playback rate should be changed in sign');
        assert_equals(player.playState, 'running', 'AnimationPlayer should not be paused');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.reverse() reverses paused player and unpauses it');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(t));
    player.playbackRate = -0.4;
    player.pause();
    player.reverse();

    player.ready.then(t.step_func(function() {
        assert_equals(player.playbackRate, 0.4, 'Playback rate should be changed in sign');
        assert_equals(player.playState, 'running', 'AnimationPlayer should not be paused');
        t.done();
    }), t.unreached_func('Current ready promise should not be rejected'));
}, 'AnimationPlayer.reverse() reverses paused player and unpauses it');


// 06/11/2014 finished here

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    var timesToTest = [
        1,
        100,
        player.source.endTime / 2,
        player.source.endTime - 100,
        player.source.endTime - 1];

    timesToTest.forEach(function(time) {
        player.currentTime = time;
        player.pause();
        var playbackRate = player.playbackRate; 
        player.reverse();

        assert_equals(player.playbackRate, -1 * playbackRate, 'Playback rate should be ' +
            'changed in sign');
        assert_false(player.paused, 'AnimationPlayer should not be paused');
    });
    testFinishedFalse(t, player);
}, 'AnimationPlayer.reverse() reverses player when player is not limited');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.pause();
    player.playbackRate = 0;

    var timesToTest = [
        -100,
        -1,
        0,
        1,
        100,
        player.source.endTime / 2,
        player.source.endTime - 100,
        player.source.endTime - 1,
        player.source.endTime,
        player.source.endTime + 1,
        player.source.endTime + 100];

    timesToTest.forEach(function(time) {
        player.currentTime = time;
        player.reverse();

        assert_equals(player.currentTime, time, 'Seek should not be performed');
        assert_true(player.paused, 'AnimationPlayer paused state should not be changed');
    });
}, 'Test that if playback rate is zero reverse() call should do nothing');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = player.source.endTime + 100;
    player.reverse();

    assert_equals(player.currentTime, player.source.endTime,
        'Seek to source content end time should be performed');
}, 'Test that if playback rate is > 0 and effective current time > source ' +
    'content\'s end time then seek to source end time');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -1;
    player.currentTime = -100;
    player.reverse();

    assert_equals(player.currentTime, 0, 'Seek to zero time should be performed');
}, 'Test that if playback rate is < 0 and effective current time < 0 then seek to 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    setTimeout(function() {
        t.step(function() {
            player.reverse();
            setTimeout(function() {
                t.step(function() {
                    assert_approx_equals(player.currentTime, 0, EPSILON, 'Time should go ' +
                        'backward after reverse() call');
                    t.done();
                });
            }, 50);
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = player.source.endTime + 100;
    player.reverse();
    setTimeout(function() {
        t.step(function() {
            assert_approx_equals(player.currentTime, player.source.endTime - 50, EPSILON,
                'Time should go backward after reverse() call');
            t.done();
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time when current time > end time');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = player.source.endTime / 2;
    player.playbackRate = -1;
    setTimeout(function() {
        t.step(function() {
            player.reverse();
            setTimeout(function() {
                t.step(function() {
                    assert_approx_equals(player.currentTime, player.source.endTime / 2, EPSILON,
                        'Time should go backward after reverse() call');
                    t.done();
                });
            }, 50);
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time when playback rate < 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = -100;
    player.playbackRate = -1;
    player.reverse();
    setTimeout(function() {
        t.step(function() {
            assert_approx_equals(player.currentTime, 50, EPSILON,
                'Time should go backward after reverse() call');
            t.done();
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time when playback rate < 0 and ' +
    'current time < 0');
</script>
</body>
